int -> long, ahaha

+++ src/main/java/ru/hse/mmstr_project/se/kafka/dto/SenderRequestDto.java
@@ -9,6 +9,6 @@ public record SenderRequestDto(
         String telegramId,
         List<String> wayToNotify,
         String phoneNumber,
-        Integer chatId,
+        Long chatId,
         String email) {
 }
